refactor: improve log details view
